architecture: Add simulate_instruction_fixup()
A few architectures (gfx950, gfx11, gfx12) have their need and way
of raising a trap exception after executing an instruction.  So far,
they have been doing it in an ad-hoc manner.  This patch adds
a pure virtual method to the "amdgcn_architecutre_t" and let each
architecture set up the necessary bits to raise the exception.
As a result, the "simulate_instruction ()" method becomes simpler.

Change-Id: I6e453ad78554eb1f2182bc56f782ae0148070bc9