libhsakmt: Fix struct queue flexible array member ordering (#243)
The cu_mask[0] flexible array member was incorrectly positioned before
the queue_type field in struct queue, violating C standard requirements
that flexible array members must be the last field in a structure.

queue_type field was added as part of a patch which was specific to
emulation only. Both these patches need not be upstreamed.

DO NOT UPSTREAM

Signed-off-by: Sreekant Somasekharan <Sreekant.Somasekharan@amd.com>